chore: rename tools/install in debug migration
