+++ packages/fx-core/src/core/middleware/utils/debug/taskMigrator.ts
@@ -5,7 +5,7 @@ import { CommentArray, CommentJSONValue, parse } from "comment-json";
 import { DebugMigrationContext } from "./debugMigrationContext";
 import { Prerequisite, TaskCommand } from "../../../../common/local";
 import { isCommentArray, isCommentObject } from "./debugV3MigrationUtils";
-import { InstallToolArgs } from "../../../../component/driver/tools/interfaces/InstallToolArgs";
+import { InstallToolArgs } from "../../../../component/driver/prerequisite/interfaces/InstallToolArgs";
 import { BuildArgs } from "../../../../component/driver/interface/buildAndDeployArgs";
 
 export function migrateTransparentPrerequisite(context: DebugMigrationContext): void {
