+++ packages/fx-core/templates/core/v3Migration/js.ts.app.local.yml
@@ -3,7 +3,7 @@ version: 1.0.0
 {{#config.deploy}}
 deploy:
   {{#tools}}
-  - uses: tools/install 
+  - uses: prerequisite/install 
     with:
       {{#devCert}}
       devCert:
