+++ packages/fx-core/tests/core/middleware/testAssets/v3Migration/debug/transparent-tab/expected/app.local.yml
@@ -1,7 +1,7 @@
 version: 1.0.0
 
 deploy:
-  - uses: tools/install 
+  - uses: prerequisite/install 
     with:
       devCert:
         trust: true
